+++ runner/android_test_orchestrator/CHANGELOG.md
@@ -6,6 +6,10 @@
 
 **Bug Fixes**
 
+* Fix a bug where the instrumentation test application would not startup if the
+arguments passed to `ORCHESTRATOR_FORWARDED_INSTRUMENTATION_ARGS` contains
+spaces.
+
 **New Features**
 
 **Breaking Changes**